change folder hierarchy
폴더 구조 대폭 수정 및 정리

+++ simulation/2D/header_background/drawgrid.h
@@ -8,12 +8,12 @@ using namespace std;
 
 void make_Points_of_grids( vector<Vector2D> &grid, int n ) {
 
-	float j = 1;
+	double j = 1;
 	//가로줄
 	for (int i = 0; i < n-1; i++) {
 		
-		grid[2 * i] = Vector2D(0, 1 /(float)n * (float)j);
-		grid[2 * i + 1] = Vector2D(1, 1 / (float)n * (float)j);
+		grid[2 * i] = Vector2D(0, 1 /(double)n * (double)j);
+		grid[2 * i + 1] = Vector2D(1, 1 / (double)n * (double)j);
 		j = j + 1;
 		
 	}
@@ -22,8 +22,8 @@ void make_Points_of_grids( vector<Vector2D> &grid, int n ) {
 
 	//세로줄
 	for (int i = n - 1; i < 2 * (n - 1); i++) {
-		grid[2 * i] = Vector2D(1 / (float)n * (float)j, 0);
-		grid[2 * i + 1] = Vector2D(1 / (float)n * (float)j,1);
+		grid[2 * i] = Vector2D(1 / (double)n * (double)j, 0);
+		grid[2 * i + 1] = Vector2D(1 / (double)n * (double)j,1);
 		j = j + 1;
 	}
